chore: implement soft-deletes for games and reviews
* chore: implement soft delete for games-reviews

* fix: fix soft delete errors

* chore: use exceptionmappers for report controller exceptions

* Merge branch 'chore/rest' into chore/soft_deletes

* fix: use exceptionmappers for report controller exceptions

* fix: hide 'resolve' link when not moderator

* Merge branch 'chore/rest' into chore/soft_deletes

* chore: merge chore/rest

* fix: show resolve link

* fix: add missing transactional decorator


Approved-by: frupnik

+++ persistence/src/main/java/ar/edu/itba/paw/persistence/ReviewHibernateDao.java
@@ -8,6 +8,7 @@ import ar.edu.itba.paw.dtos.ordering.ReviewOrderCriteria;
 import ar.edu.itba.paw.enums.Difficulty;
 import ar.edu.itba.paw.enums.FeedbackType;
 import ar.edu.itba.paw.enums.Platform;
+import ar.edu.itba.paw.enums.ReportState;
 import ar.edu.itba.paw.models.*;
 import ar.edu.itba.paw.models.keys.ReviewFeedbackId;
 import ar.edu.itba.paw.persistence.helpers.DaoUtils;
@@ -49,6 +50,7 @@ public class ReviewHibernateDao implements ReviewDao, PaginationDao<ReviewFilter
     @Override
     public Optional<Review> findById(long id, Long activeUserId) {
         Review review = em.find(Review.class, id);
+        if(review == null || review.getDeleted()) return Optional.empty();
         User user = em.find(User.class, activeUserId != null ? activeUserId : -1);
         if (user != null) {
             ReviewFeedback reviewFeedback = em.find(ReviewFeedback.class, new ReviewFeedbackId(user, review));
@@ -56,13 +58,13 @@ public class ReviewHibernateDao implements ReviewDao, PaginationDao<ReviewFilter
                 review.setFeedback(reviewFeedback.getFeedback());
             }
         }
-        return Optional.ofNullable(review);
+        return Optional.of(review);
     }
 
     @Override
     public Optional<Review> update(long id, SaveReviewDTO reviewDTO) {
         Review review = em.find(Review.class, id);
-        if (review == null) {
+        if (review == null || review.getDeleted()) {
             return Optional.empty();
         }
         if (reviewDTO.getTitle() != null) {
@@ -112,7 +114,8 @@ public class ReviewHibernateDao implements ReviewDao, PaginationDao<ReviewFilter
                 .withList("r.authorid", filter.getAuthors())
                 .withExact("r.gameid", filter.getGameId())
                 .withExact("r.completed", filter.getCompleted())
-                .withExact("r.replayability", filter.getReplayable());
+                .withExact("r.replayability", filter.getReplayable())
+                .withExact("r.deleted", filter.getDeleted());
         if (filter.getMinTimePlayed() != null) {
             queryBuilder = queryBuilder.withGreaterOrEqual("COALESCE(r.gamelength, 0)", 3600 * filter.getMinTimePlayed());
         }
@@ -181,14 +184,27 @@ public class ReviewHibernateDao implements ReviewDao, PaginationDao<ReviewFilter
     @Override
     public boolean deleteReview(long id) {
         Review review = em.find(Review.class, id);
-        if(review == null) return false;
-        em.remove(review);
+        if(review == null || review.getDeleted()) return false;
+        review.setDeleted(true);
         return true;
     }
 
     @Override
+    public int deleteReviewsOfGame(long gameId) {
+        Game game = em.find(Game.class, gameId);
+        if (game == null || game.getDeleted()) return 0;
+        final TypedQuery<Review> query = em.createQuery("update Review set deleted = true where reviewedGame = :game", Review.class);
+        query.setParameter("game", game);
+        return query.executeUpdate();
+    }
+
+    @Override
     public Optional<FeedbackType> getReviewFeedback(long reviewId, long userId) {
-        ReviewFeedback feedback = em.find(ReviewFeedback.class, new ReviewFeedbackId(em.find(User.class, userId), em.find(Review.class, reviewId)));
+        User user = em.find(User.class, userId);
+        Review review = em.find(Review.class, reviewId);
+        if(review == null || review.getDeleted())
+            return Optional.empty();
+        ReviewFeedback feedback = em.find(ReviewFeedback.class, new ReviewFeedbackId(user, review));
         if(feedback == null) return Optional.empty();
         return Optional.of(feedback.getFeedback());
     }
@@ -197,7 +213,7 @@ public class ReviewHibernateDao implements ReviewDao, PaginationDao<ReviewFilter
     public Optional<ReviewFeedback> editReviewFeedback(long reviewId, long userId, FeedbackType oldFeedback, FeedbackType feedback) {
         Review review = em.find(Review.class, reviewId);
         ReviewFeedback reviewFeedback = em.find(ReviewFeedback.class, new ReviewFeedbackId(em.find(User.class, userId), review));
-        if(reviewFeedback == null || review == null) {
+        if(reviewFeedback == null || review == null || review.getDeleted()) {
             return Optional.empty();
         }
         if (oldFeedback != feedback) {
@@ -214,7 +230,7 @@ public class ReviewHibernateDao implements ReviewDao, PaginationDao<ReviewFilter
     public Optional<ReviewFeedback> addReviewFeedback(long reviewId, long userId, FeedbackType feedback) {
         Review likedReview = em.find(Review.class, reviewId);
         User likedUser = em.find(User.class, userId);
-        if (likedReview == null || likedUser == null) return Optional.empty();
+        if (likedReview == null || likedReview.getDeleted() || likedUser == null) return Optional.empty();
         ReviewFeedback possibleFeedback = em.find(ReviewFeedback.class, new ReviewFeedbackId(likedUser, likedReview));
         if (possibleFeedback != null) return Optional.empty();
         ReviewFeedback reviewFeedback = new ReviewFeedback(likedUser, likedReview, feedback);
@@ -233,7 +249,7 @@ public class ReviewHibernateDao implements ReviewDao, PaginationDao<ReviewFilter
     public boolean deleteReviewFeedback(long reviewId, long userId, FeedbackType oldFeedback) {
         Review likedReview = em.find(Review.class, reviewId);
         User likedUser = em.find(User.class, userId);
-        if (likedReview == null || likedUser == null) return false;
+        if (likedReview == null || likedReview.getDeleted() || likedUser == null) return false;
         ReviewFeedback possibleFeedback = em.find(ReviewFeedback.class, new ReviewFeedbackId(likedUser, likedReview));
         if (possibleFeedback == null) return false;
         em.remove(possibleFeedback);
